Return 404 when TransformSession doesn't have enough data.
